# openapi.yaml
openapi: 3.0.3
info:
  title: Database Query Management API
  description: |
    A comprehensive API for managing database query requests with approval workflow.
    
    ## Features
    - **Multi-Database Support**: PostgreSQL and MongoDB
    - **Approval Workflow**: Manager approval system for database operations
    - **Audit Trail**: Complete logging of all database operations
    - **Security-First**: Built-in validation and access control
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token.
    
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://zluri-bootcamp-backend.up.railway.app/api/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /request:
    post:
      tags:
        - Request Management
      summary: Submit query request
      description: |
        Submit a new database query or script request for approval.
        
        ## Request Types
        - **Query**: Direct SQL/MongoDB query execution
        - **Script**: JavaScript code with database operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/QueryRequest'
                - $ref: '#/components/schemas/ScriptRequest'
            examples:
              postgres_query:
                summary: PostgreSQL Query Request
                value:
                  databasetype: "POSTGRES"
                  instanceid: "1"
                  dbname: "postgres"
                  query: "SELECT * FROM users WHERE active = true LIMIT 10;"
                  comments: "Need to check active user data"
                  podid: "pod-1"
              postgres_script:
                summary: PostgreSQL Script Request
                value:
                  databasetype: "POSTGRES"
                  instanceid: "1"
                  dbname: "postgres"
                  script: "async function getUsers() { const result = await query('SELECT * FROM users'); console.log('Users:', result.rows.length); }"
                  comments: "Execute user analysis script"
                  podid: "pod-1"
              mongo_query:
                summary: MongoDB Query Request
                value:
                  databasetype: "MONGO"
                  instanceid: "2"
                  dbname: "test_mongo"
                  query: "db.users.find({status: 'active'})"
                  comments: "Fetch active users"
                  podid: "pod-1"
              dangerous_query:
                summary: High-Risk Query (will be flagged)
                value:
                  databasetype: "POSTGRES"
                  instanceid: "1"
                  dbname: "postgres"
                  query: "DROP TABLE users;"
                  comments: "Clean up old data"
                  podid: "pod-1"
      responses:
        '200':
          description: Request submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestSubmissionResponse'
              examples:
                success:
                  summary: Request submitted
                  value:
                    success: true
                    req_id: "123"
                    status: "PENDING"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /requests/{reqid}/result:
    get:
      tags:
        - Request Management
      summary: Get request execution result
      description: Retrieve the execution result for a specific request
      parameters:
        - name: reqid
          in: path
          required: true
          schema:
            type: string
          description: Request ID
          example: "123"
      responses:
        '200':
          description: Execution result retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResultResponse'
              examples:
                success:
                  summary: Successful execution
                  value:
                    success: true
                    output: "Query executed successfully. 10 rows returned."
                    response_time: 150
                    status: "success"
                    error: null
                failure:
                  summary: Failed execution
                  value:
                    success: true
                    output: null
                    response_time: 50
                    status: "failure"
                    error: "Syntax error in SQL query"
                pending:
                  summary: Not executed yet
                  value:
                    success: true
                    output: null
                    response_time: null
                    status: "pending"
                    message: "Request not yet executed"
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Approval Management Endpoints
...
  /approvals/{reqId}/action:
    post:
      tags:
        - Approval Management
      summary: Approve or reject request
      description: |
        Approve or reject a pending request.
      parameters:
        - name: reqId
          in: path
          required: true
          schema:
            type: string
          description: Request ID
          example: "123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalActionRequest'
            examples:
              approve:
                summary: Approve request
                value:
                  action: "approve"
              reject:
                summary: Reject request
                value:
                  action: "reject"
                  reason: "Query needs modification before approval"
      responses:
        '200':
          description: Action completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalActionResponse'
              examples:
                approved:
                  summary: Request approved
                  value:
                    success: true
                    status: "approval"
                rejected:
                  summary: Request rejected
                  value:
                    success: true
                    status: "reject"
                    reason: "Query needs modification before approval"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    QueryRequest:
      type: object
      required:
        - databasetype
        - instanceid
        - dbname
        - query
        - podid
      properties:
        databasetype:
          type: string
          enum: [POSTGRES, MONGO]
          description: Database type
          example: "POSTGRES"
        instanceid:
          type: string
          description: Database instance ID
          example: "1"
        dbname:
          type: string
          description: Database name
          example: "postgres"
        query:
          type: string
          description: SQL or MongoDB query to execute
          example: "SELECT * FROM users LIMIT 10;"
        comments:
          type: string
          description: Optional comments about the request
          example: "Need to check user data"
        podid:
          type: string
          description: Pod ID for approval routing
          example: "pod-1"
    ScriptRequest:
      type: object
      required:
        - databasetype
        - instanceid
        - dbname
        - script
        - podid
      properties:
        databasetype:
          type: string
          enum: [POSTGRES, MONGO]
          description: Database type
          example: "POSTGRES"
        instanceid:
          type: string
          description: Database instance ID
          example: "1"
        dbname:
          type: string
          description: Database name
          example: "postgres"
        script:
          type: string
          description: JavaScript code to execute
          example: "async function getUsers() { const result = await query('SELECT * FROM users'); console.log('Users:', result.rows.length); }"
        comments:
          type: string
          description: Optional comments about the request
          example: "Execute user analysis script"
        podid:
          type: string
          description: Pod ID for approval routing
          example: "pod-1"
    RequestSubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        req_id:
          type: string
          description: Generated request ID
          example: "123"
        status:
          type: string
          enum: [PENDING]
          example: "PENDING"
...
    ExecutionResultResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        output:
          type: string
          nullable: true
        response_time:
          type: integer
          nullable: true
        status:
          type: string
          enum: [success, failure, pending]
        error:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
          description: Additional status message

    # Approval Management Schemas
...
    ApprovalActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [approve, reject]
          description: Action to perform
          example: "approve"
        reason:
          type: string
          description: Reason for rejection (required if action is reject)
          example: "Query needs modification before approval"
    ApprovalActionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: string
          enum: [approval, reject]
          example: "approval"
        reason:
          type: string
          nullable: true
          description: Rejection reason (only present if rejected)

    # Common Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error description
          example: "Invalid request parameters"
        error:
          type: string
          nullable: true
          description: Detailed error information
...